% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{est}
\alias{est}
\alias{est_simple}
\alias{est_mand}
\alias{est_vol}
\title{Estimate total harvest}
\usage{
est_mand(simdat)

est_simple(simdat, poststrat = FALSE)

est_vol(simdat, poststrat = FALSE)
}
\arguments{
\item{simdat}{Simulation data. An output from either \code{pop_simple()},
\code{pop_mand()}, or \code{pop_vol()}.}

\item{poststrat}{Logical. If \code{TRUE}, post-stratify respondents by the
\code{group} variable. If a follow up survey is simulated, it will also be
post-stratified.}
}
\value{
A tibble, containing an estimate of harvest for each level of
response rate, standard error, and absolute percent error, among
other important metadata.
}
\description{
Estimates total harvest from a dataset created by one of
the explicit \code{\link{pop}} functions.
\itemize{
\item{\code{est_mand()} creates harvest estimates from a
\code{pop_mand()} output.}
\item{\code{est_simple()} creates harvest estimates from a
\code{pop_simple()} output.}
\item{\code{est_vol()} creates harvest estimates from a
\code{pop_vol()} output.}
}
}
\details{
These functions rely on the structure of \code{simdat} being unchanged from
its original structure to know how to create estimates.
If you must manipulate the dataset, create estimates
from it first, then manipulate it.\cr\cr
If the argument \code{simdat} contains data with multiple response rates
simulated, the \code{est} functions will create a different estimate for
each unique level of response rate within that data.\cr\cr
All \code{est} functions, except
\code{est_mand()}, use \code{survey::svydesign()} and
\code{survey::svytotal()} to estimate total harvest. Estimates are
calculated a bit different when a follow up survey is simulated in the
data to be estimated. In that case, two seperate estimates are made using
\code{survey::svytotal()}. One estimate from the respondents to
the initial survey and the other from respondents
to follow up surveys. In the case of \code{est_simple()}, the proportion of
(hunters initially responding / hunters sampled) is used to scale the initial
estimate (i.e. the prop. of respondents is assumed to make up the same
prop. of the entire population.) Then, the follow up respondents are assumed
to be representative of the remaining proportion of the entire population.
Both scaled estimates are then added together to create the combined
estimate. This is similar to \code{est_vol()}, except in that case, the
proportion of respondents \emph{is} the proportion of the population
that responded.\cr\cr
\code{est_mand()} assumes 100\% reporting by successful hunters,
so it simply sums up the total reported harvest.\cr\cr
If \code{poststrat = TRUE}, \code{survey::postStratify()} is used,
with the \code{partial} argument set to \code{TRUE}.
}
\examples{
# Estimate harvest from a simulation where hunters report voluntarily:
# first, you must create the data:
dat <- hss::pop_vol(
  n          = 1000,
  split      = 0.8,
  success1   = 0.3,
  success0   = 0.7,
  resp       = c(0.2, 0.6, 0.8),
  resp_bias  = 1.2
)

# then you can create estimates:
dat_ests <- est_vol(dat)

# if you want to post-stratify by the "group" variable:
dat_ests_ps <- est_vol(dat, poststrat = TRUE)
}
\seealso{
\code{\link{pop}} for how to create the data to be input to this
function.\cr\cr
\code{\link{pae}} if you do not need the raw population data and want to
skip this intermediate step and get straight to estimates. \cr\cr
\code{\link[survey]{svydesign}}, \code{\link[survey]{svytotal}}, and
\code{\link[survey]{postStratify}} for
details on how most estimates and standard errors are calculated.
}
