% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\alias{pop_simple}
\alias{pop_mand}
\alias{pop_vol}
\title{Population and harvest survey simulations}
\usage{
pop(n, split = 1, success1, success0 = success1)

pop_mand(n, split = 1, success1, success0 = success1, resp)

pop_simple(
  n,
  split = 1,
  success1,
  success0 = success1,
  sample,
  resp,
  resp_bias,
  fus = FALSE,
  fus_scale = NULL
)

pop_vol(
  n,
  split = 1,
  success1,
  success0 = success1,
  resp,
  resp_bias,
  fus = FALSE,
  fus_sample = NULL,
  fus_scale = NULL
)
}
\arguments{
\item{n}{The desired population size.}

\item{split}{Proportion of the population that is placed into group 1.
The remaining proportion (i.e. \code{1 - split}) will be in group 0.}

\item{success1}{Probability of a hunter in group 1 to harvest}

\item{success0}{Probability of a hunter in group 0 to harvest}

\item{resp}{Probability/probabilities of response. Can be a vector of
length > 1.
\itemize{
\item In \code{pop_simple()} and \code{pop_vol()} it defines response
probabilities for unsuccessful hunters.
\item In \code{pop_mand()} it defines
response probabilities for successful hunters, as they are the only ones
mandated to report.
}}

\item{sample}{Probability a hunter is sampled for a survey}

\item{resp_bias}{Scales \code{resp} to create probabilities of response
for successful hunters. Introduces response bias between successful and
unsuccessful hunters if not equal to one.}

\item{fus}{Logical. If TRUE, a single follow up survey will be simulated.}

\item{fus_scale}{Scales initial response probabilities,
creating new probabilities of response for follow up surveys.}

\item{fus_sample}{Probability that a non-respondent is sampled for a follow
up survey.}
}
\value{
A tibble, containing the simulated population and response results,
alongside other metadata.
}
\description{
The pop functions will simulate a population of hunters and whether or not
they were successful in harvesting. If a survey method is specified,
response to surveys are also simulated. \cr
\itemize{
\item \code{pop()} creates a population of hunters and simulates
harvest.
\item \code{pop_simple()} creates a simulation where a population of hunters
are surveyed using a simple random sample.
\item \code{pop_mand()} creates a simulation where only
successful hunters are mandated to report.
\item \code{pop_vol()} creates a simulation where
reporting is voluntary, and both successful and unsuccessful hunters report.
Follow up surveys are completed by simple random sampling.
}
}
\details{
If any scaling arguments scale probabilities to be > 1, the probabilities
will silently be changed to 1. \cr\cr
If argument \code{resp} is defined to be a vector with length > 1, the same
population will be recycled and put through a different response simulation
for each element within the vector.
}
\examples{
# Simulate a dataset that contains a population of 1,000 hunters and
# simulates harvest where each hunter has a probability of 0.3 to harvest:
pop(n = 1000, success1 = 0.3)

# A similar population as above example, but with a simulation where
# successful hunters are mandated to report. In this case, only ~70\% of them
# actually report:
pop_mand(
  n          = 1000,
  success1   = 0.3,
  resp       = 0.7,
)

# A population where ~70\% of hunters are in group 1 and
# harvest with a probability of 0.25, while the remaining ~30\%
# are in group 0 and harvest with a probability of 0.4. Hunters are
# sampled by simple random sampling  at a probability of 0.5. Simulate
# multiple levels of response, where successful hunters are 1.2 times
# more likely to respond than unsuccessful hunters:
pop_simple(
  n          = 1000,
  split      = 0.7,
  success1   = 0.25,
  success0   = 0.4,
  sample     = 0.5,
  resp       = seq(0.3, 0.8, 0.1),
  resp_bias  = 1.2
)

# Simulate a similar population as the one in the example above.
# Except this time complete a follow up survey of the
# non-respondents, where hunters are only 0.7 times as likely
# (i.e. ~30\% less likely) to respond to the follow up survey as they
# were to the initial survey:
pop_simple(
  n          = 1000,
  split      = 0.7,
  success1   = 0.4,
  success0   = 0.25,
  sample     = 0.5,
  resp       = seq(0.3, 0.8, 0.1),
  resp_bias  = 1.2,
  fus        = TRUE,
  fus_scale  = 0.7
)

}
