% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{Estimate total harvest}
\usage{
est(simdat)
}
\arguments{
\item{simdat}{Simulation data. An output from either \code{pop_simple()},
\code{pop_mand()}, or \code{pop_vol()}.}
}
\value{
A tibble, containing an estimate of harvest for each level of
response rate and response bias. Also reports standard error, and mean
absolute percent error, among other metadata.
}
\description{
Estimates total harvest from a data set created by one of
the \code{\link{survey}} functions.
}
\details{
This function relies on the structure of \code{simdat} being unchanged from
its original structure to know how to create estimates.
If you must manipulate the data set, create estimates
from it first, then manipulate it.\cr\cr
If estimates are to be made from a data set that contains multiple response
simulations, multiple response biases, and/or multiple response rates,
this function breaks the data set down into single populations and creates
an estimate for every single population. It then returns the averages for
these estimates.\cr\cr
All estimations, except those made from \code{pop_mand()} outputs,
use \code{survey::svydesign()} and
\code{survey::svytotal()} to estimate total harvest. Estimates are
calculated a bit different when a follow up survey is simulated in the
data. In that case, two separate estimates are made using
\code{survey::svytotal()}. One estimate from the respondents to
the initial survey and the other from respondents
to follow up surveys. In the case of estimates from \code{pop_simple()},
the proportion of (hunters initially responding / hunters sampled) is
used to scale the initial estimate (i.e. the prop. of respondents is
assumed to make up the same
prop. of the entire population). Then, the follow up respondents are assumed
to be representative of the remaining proportion of the entire population.
This proportion is used to scale the follow up estimate.
Both scaled estimates are then added together to create the combined
estimate. This is similar to estimates made from \code{pop_vol()},
except in that case, the proportion of initial respondents \emph{is} the
proportion of the population that responded.\cr\cr
\code{est_mand()} assumes 100\% reporting by successful hunters,
so it simply sums up the total reported harvest.\cr\cr
}
\examples{
# Estimate harvest from a simulation where hunters report voluntarily:
# first, you must create the data:
dat <- pop_vol(
  n          = 1000,
  split      = 0.8,
  success1   = 0.3,
  success0   = 0.7,
  resp       = c(0.2, 0.6, 0.8),
  bias       = 1.2,
  times      = 50
)

# Then you can create estimates:
dat_ests <- est(dat)


# Usage is the same for all possible inputs of "simdat".

}
\seealso{
\code{\link{pop}} for how to create the data to be input to this
function.\cr\cr
\code{\link[survey]{svydesign}}, \code{\link[survey]{svytotal}} for
details on how estimates are calculated.
}
