% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est}
\alias{est}
\title{Estimate Total Harvest}
\usage{
est(simdat)
}
\arguments{
\item{simdat}{Survey simulation data. Must be an object
created from one of the three \code{\link{survey}} functions.}
}
\value{
A data frame, containing the following variables:
\itemize{
\item \code{N}: Hunter population size.
\item \code{resp_bias}: Bias of successful hunters to report,
    relative to unsuccessful hunters.
\item \code{resp_rate}: Underlying response probability, before any
    scaling/bias.
\item \code{mean_init_n}: Average total initial survey responses.
\item \code{mean_fol_n}: Average total follow-up survey responses.
\item \code{true_harvest}: True harvest of population.
\item \code{min_hvst_est}: Minimum harvest estimate of the survey
    repetitions (if any) at the response bias and response rate shown for that row.
\item \code{max_hvst_est}: Maximum harvest estimate.
\item \code{mean_hvst_est}: Average harvest estimate.
\item \code{mean_SE}: Average standard error across repetitions.
\item \code{MARE}: Mean absolute relative error.
\item \code{RRMSE}: Relative root mean squared error.
}
}
\description{
A generic function for estimating total harvest from objects
created by the \code{\link{survey}} simulation functions.
}
\details{
In simulations with follow up surveys, two separate estimates are made.
One estimate from the respondents to
the initial survey and the other from respondents to follow up surveys.
In the case of estimates from \code{simple()} outputs,
the proportion of \code{init_resp/init_sample} is
used to scale the initial estimate (i.e. the prop. of total respondents
to total sampled is assumed to make up the same
prop. of the entire population). Then, the follow up respondents are assumed
to be representative of the remaining proportion of the entire population.
This remaining proportion is used to scale the follow up estimate.
Both scaled estimates are then added together to create the combined
total population estimate.\cr

This is similar to estimates made from \code{census()} outputs,
except in that case the initial scaling proportion is \code{init_resp/N}.

Estimates for \code{mand()} outputs assume 100\% reporting by successful
hunters for initial reports if there is no follow up. If a follow up survey
was simulated, it creates a harvest estimate from the follow up sample to
estimate total harvest for \emph{the non-reporting portion of
the population only}, and then adds that to the sum of
initially reported harvests.\cr\cr
}
\examples{
# First, make a population:
my_pop <- pop(N = 10000, split = 0.7, success1 = 0.25, success2 = 0.32)

# Then simulate a survey for that population:
my_pop_simple_followup <- simple(
  x = my_pop,
  sample = 0.5,
  resp = c(0.4, 0.6),
  bias = seq(1, 1.3, 0.1),
  fol = TRUE,
  fol_scale = 0.7,
  times = 10
)

# Finally, make your estimates:
est(my_pop_simple_followup)
}
\seealso{
\code{\link{pop}} and \code{\link{survey}} for how to create the data to be
input to this function.\cr\cr
}
