% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hss_documentation.R
\docType{package}
\name{hss}
\alias{hss}
\title{hss: Hunter survey simulations, and estimations of harvest.}
\description{
There are three types of functions within this package:
\code{\link{pop}},
\code{\link{est}},
and \code{\link{pae}}
}
\section{pop functions}{

Simulate a population of hunters, simulate harvest, and
their response to different survey methods.
}

\section{est function}{

Make total harvest estimates from simulated data created by
pop functions.
}

\section{pae functions}{

Simulate populations, and create estimates from them, all in one function.
Useful when the population data is not needed.
}

