% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Population and harvest survey simulations}
\usage{
pop(n, split = 1, success1, success0 = success1)
}
\arguments{
\item{n}{The desired population size.}

\item{split}{Proportion of the population that is placed into group 1.
The remaining proportion (i.e. \code{1 - split}) will be in group 0.}

\item{success1}{Probability of a hunter in group 1 to harvest}

\item{success0}{Probability of a hunter in group 0 to harvest}
}
\value{
A tibble that contains the variables:
\itemize{
\item \code{pop_size}: The population size.
\item \code{true_harvest}: The sum of harvests from the population.
\item \code{group}: The group in which the hunter was placed.
\item \code{harvest}: "1" for a successful hunter, and "0" if unsuccessful.
}
}
\description{
The pop function simulates a population of hunters and whether or not
they were successful in harvesting.
}
\examples{
# Simulate a population of 1,000 hunters and
# simulates harvest where each hunter has a probability of 0.3 to harvest:
pop(n = 1000, success1 = 0.3)

}
